[dtensor] group common rules together
This PR group common rules together, to allow each op category reuse
those rules easily
