+++ spmd/tensor/ops/matrix_ops.py
@@ -1,8 +1,7 @@
 # Copyright (c) Meta Platforms, Inc. and affiliates
 # implement matrix related ops for distributed tensor
 from spmd.tensor.dispatch import OpSchema, OutputSharding
-from spmd.tensor.ops.math_ops import einop_rule
-from spmd.tensor.ops.pointwise_ops import pointwise_rule
+from spmd.tensor.ops.common_rules import einop_rule, pointwise_rule
 from spmd.tensor.ops.utils import register_prop_rule
